Login done

+++ app/src/main/java/com/example/utente_pc1/provamvvm/ui/login/LoginActivity.java
@@ -6,6 +6,7 @@ import android.arch.lifecycle.LifecycleOwner;
 import android.arch.lifecycle.Observer;
 import android.content.Context;
 import android.content.Intent;
+import android.os.AsyncTask;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 
@@ -14,6 +15,7 @@ import android.view.View;
 
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
 
 
 import com.example.utente_pc1.provamvvm.LecApplication;
@@ -21,10 +23,13 @@ import com.example.utente_pc1.provamvvm.R;
 import com.example.utente_pc1.provamvvm.model.data.local.SingleSubj;
 import com.example.utente_pc1.provamvvm.ui.list.Listactivity;
 
+import com.example.utente_pc1.provamvvm.util.exceptions.ConnectionException;
+import com.example.utente_pc1.provamvvm.util.exceptions.LoginException;
 import com.example.utente_pc1.provamvvm.viewmodel.CustomViewModelFactory;
 import com.example.utente_pc1.provamvvm.viewmodel.ListItemViewModel;
 
 import java.util.List;
+
 
 import javax.inject.Inject;
 
@@ -43,7 +48,6 @@ public class LoginActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setTheme(R.style.AppTheme);
         setContentView(R.layout.activity_login);
-
         ((LecApplication) getApplication())
                 .getLecComponent()
                 .inject(this);
@@ -59,29 +63,68 @@ public class LoginActivity extends AppCompatActivity {
         login.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                progressDialog.show();
-                listItemViewModel.deleteAllsubj();
-                listItemViewModel.logIn(nametext.getText().toString(), password.getText().toString());
-                listItemViewModel.getSingleSubj().observe(context, new Observer<List<SingleSubj>>() {
-                    @Override
-                    public void onChanged(@Nullable List<SingleSubj> singleSubjs) {
-                        if (singleSubjs != null) {
-                            for (SingleSubj single : singleSubjs) {
-                                listItemViewModel.insertSingleSubj(single);
+                if (nametext.getText().toString().isEmpty() || password.getText().toString().isEmpty()) {
+                    Toast.makeText(getApplicationContext(), "You have to complete login form!", Toast.LENGTH_LONG)
+                            .show();
+                } else {
+                    progressDialog.show();
+                    listItemViewModel.deleteAllsubj();
+                    AsyncTask<Void, Void, Void> login = new AsyncTask<Void, Void, Void>() {
+                        private String errormessage;
+
+                        @Override
+                        protected Void doInBackground(Void... voids) {
+                            try {
+                                listItemViewModel.logIn(nametext.getText().toString(), password.getText().toString());
+                            } catch (ConnectionException e) {
+                                errormessage = e.getMessage();
+                            } catch (LoginException e1) {
+                                errormessage = e1.getMessage();
+                            }
+                            return null;
+                        }
+
+                        @Override
+                        protected void onPostExecute(Void aVoid) {
+                            if (errormessage == null) {
+                                doafterlogin();
+                            } else {
+                                progressDialog.dismiss();
+                                Toast.makeText((Context) context, errormessage, Toast.LENGTH_LONG).show();
                             }
                         }
-                        progressDialog.dismiss();
-                        Intent i = new Intent((Context) context, Listactivity.class);
-                        i.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                        startActivity(i);
-                        finish();
-                    }
-                });
+
+                    };
+                    login.execute();
+                }
             }
         });
-
-
     }
 
+    private void doafterlogin() {
+        listItemViewModel.getSingleSubj().observe(context, new Observer<List<SingleSubj>>() {
+            @Override
+            public void onChanged(@Nullable List<SingleSubj> singleSubjs) {
+                if (singleSubjs != null) {
+                    for (SingleSubj single : singleSubjs) {
+                        listItemViewModel.insertSingleSubj(single);
+                    }
+                    progressDialog.dismiss();
+                    toHomeActivity();
+                } else {
+                    progressDialog.dismiss();
+                    Toast.makeText((Context) context, "Internet Connection Missing!", Toast.LENGTH_LONG);
+                }
+
+            }
+        });
+    }
+
+    private void toHomeActivity() {
+        Intent i = new Intent((Context) context, Listactivity.class);
+        i.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+        startActivity(i);
+        finish();
+    }
 }
